Concentration risk

The sum of all observations within a radius from center point(s). In particular, it can be used to determine concentration risk in the context of the EU insurance regulation framework (Solvency II). The function offers an effective approach to calculate the 'standard formula' under Solvency II. The 'standard formula' under Solvency II asks companies to report their largest fire concentration in respect of the fire peril within a radius of 200m. This is the maximum gross sum insured of the set of buildings fully or partly located within this radius (Commission Delegated Regulation (EU), 2015, Article 132).

concentration(
  sub,
  full,
  value,
  lon_sub = lon,
  lat_sub = lat,
  lon_full = lon,
  lat_full = lat,
  radius = 200,
  display_progress = TRUE
)

Arguments

sub
data.frame of locations to calculate concentration risk for (target points).

full
data.frame to find the locations within radius r from locations in sub (reference locations).

value
Column with value in full.

lon_sub
Column in sub with longitude (lon is default).

lat_sub
Column in sub with latitude (lat is default).

lon_full
Column in full with longitude in full (lon is default).

lat_full
Column in full with latitude in full (lat is default).

radius
Radius (in meters) (default is 200m).

display_progress
Show progress bar (TRUE/FALSE).

Value

A data.frame equal to data.frame sub including an extra column concentration.

Details

The data.frame sub should include at least columns for longitude and latitude.

The data.frame full should include at least columns for longitude, latitude and value of interest to summarize.

References

Commission Delegated Regulation (EU) (2015). Solvency II Delegated Act 2015/35. Official Journal of the European Union, 58:124.

Examples

df <- data.frame(location = c("p1", "p2"), lon = c(6.561561, 6.561398), lat = c(53.21369, 53.21326))
concentration(df, Groningen, value = amount, radius = 100)

#>   location      lon      lat concentration
#> 1       p1 6.561561 53.21369          2055
#> 2       p2 6.561398 53.21326          2892