Retrieve historic weather data for the Netherlands

This function retrieves historic weather data collected by the official KNMI weather stations. See spatialrisk::knmi_stations for a list of the official KNMI weather stations.

Usage

knmi_historic_data(startyear, endyear)

Format

The returned data frame contains the following columns:

station = ID of measurement station;

date = Date;

FH = Hourly mean wind speed (in 0.1 m/s);

FX = Maximum wind gust (in 0.1 m/s) during the hourly division;

DR = Precipitation duration (in 0.1 hour) during the hourly division;

RH = Hourly precipitation amount (in 0.1 mm) (-1 for <0.05 mm);

city = City where the measurement station is located;

lon = Longitude of station (crs = 4326);

lat = Latitude of station (crs = 4326).

Arguments

startyear

start year for historic weather data.

endyear

end year for historic weather data.

Value

Data frame containing weather data and meta data for weather station locations.

Author

Martin Haringa

Examples

if (FALSE) { # \dontrun{
knmi_historic_data(2015, 2019)
} # }