Retrieve historic weather data for the Netherlands

This function retrieves historic weather data collected by the official KNMI weather stations. See spatialrisk::knmi_stations for a list of the official KNMI weather stations.

knmi_historic_data(startyear, endyear)

Arguments

startyear
start year for historic weather data.

endyear
end year for historic weather data.

Format

The returned data frame contains the following columns:

station = ID of measurement station;

date = Date;

FH = Hourly mean wind speed (in 0.1 m/s)

FX = Maximum wind gust (in 0.1 m/s) during the hourly division;

T = Temperature (in 0.1 degrees Celsius) at 1.50 m at the time of observation;

DR = Precipitation duration (in 0.1 hour) during the hourly division;

RH = Hourly precipitation amount (in 0.1 mm) (-1 for <0.05 mm);

city = City where the measurement station is located;

lon = Longitude of station (crs = 4326);

lat = Latitude of station (crs = 4326).

Value

Data frame containing weather data and meta data for weather station locations.

Examples

if (FALSE) {
knmi_historic_data(2015, 2019)
}